+++ src/watercss_function.ps1
@@ -17,11 +17,7 @@
         License: The MIT License (MIT) Copyright © 2019 Kognise
 
 .LINK
-    thanks:
-        Water.css: https://watercss.kognise.dev/
-        GitHub: https://github.com/kognise/water.css
-        License: The MIT License (MIT) Copyright © 2019 Kognise
-
+    
 
 #>
 function watercss {